% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{applyNesTransformation}
\alias{applyNesTransformation}
\title{Apply NES transformation}
\usage{
applyNesTransformation(x, transformation)
}
\arguments{
\item{x}{MAtrix or list of matrices}

\item{transformation}{Function generated by nesTRansformationFunction() function}
}
\value{
Transformed matrix or list of matrices
}
\description{
Apply NES transformation
}
