% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{colorScale}
\alias{colorScale}
\title{colorScale}
\usage{
colorScale(x, color = c("royalblue", "firebrick2"), gama = 1,
  alpha = 1, scmax = 0, nacol = "grey80")
}
\arguments{
\item{x}{Vector or matrix of numeric values}

\item{color}{Vector of character strings indicating the colors for the scale. Up to three colors can be defined. While is used for the missing color}

\item{gama}{Number indicating the gama transformation}

\item{alpha}{Number between 0 and 1 indicating the transparency of the color (1 for absolute color)}

\item{scmax}{Number indicating the maximum value for the scale}

\item{nacol}{Character string indicating the color for missing values}
}
\value{
Vector of colors
}
\description{
This function generates a color scale
}
