% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{colVars}
\alias{colVars}
\title{Variance of columns for arrays with NA values}
\usage{
colVars(x)
}
\arguments{
\item{x}{Numeric matrix}
}
\value{
Vector with the variance by column results
}
\description{
This function computes the variance by columns ignoring NA values
}
