% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{colWVars}
\alias{colWVars}
\title{Weighted variance by columns}
\usage{
colWVars(x, w)
}
\arguments{
\item{x}{Sparse matrix}

\item{w}{Numeric vector of weights}
}
\value{
Vector of weighted variances
}
\description{
Weighted variance by columns
}
