% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{computeNesForMatrixRow}
\alias{computeNesForMatrixRow}
\title{Compute NES for a specific row of a matrix}
\usage{
computeNesForMatrixRow(i, nesmat, nullmat)
}
\arguments{
\item{i}{Integer indicating the row number}

\item{nesmat}{Matrix of scores}

\item{nullmat}{Matrix for the null model}
}
\value{
Vctor of NES
}
\description{
This function estimates the NES for a specific row of a matrix
}
