% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{concatenateMatrixList}
\alias{concatenateMatrixList}
\title{Integrate list of matrices}
\usage{
concatenateMatrixList(x, method = c("union", "intersection"))
}
\arguments{
\item{x}{List of matrices}
}
\value{
Matrix
}
\description{
Integrate list of matrices
}
