% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{cpm}
\alias{cpm}
\title{Counts per million implementation for sparse matrices}
\usage{
cpm(x)
}
\arguments{
\item{x}{Numeric sparse matrix}
}
\value{
sparse matrix of CPM
}
\description{
Counts per million implementation for sparse matrices
}
