% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.r
\docType{data}
\name{geneLength}
\alias{geneLength}
\title{Average length of human and mouse known genes}
\format{
Vector of integers indicating the average length in bp
  for each gene, indicated with EntrezIDs as name argument.
To access this data use:
\describe{
  \item{data(hg38)}{Human}
  \item{data(mm10)}{Mouse}
}
}
\usage{
geneLength
}
\description{
A dataset containing the average length for known mouse
and human genes
}
\keyword{datasets}
