% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.r
\docType{data}
\name{genePosition}
\alias{genePosition}
\title{Chromosomal coordinate of human and mouse known genes}
\format{
data.frame with 2 columns: Chromosome and Coordinate.
To access this data use:
\describe{
  \item{data(hg38)}{Human}
  \item{data(mm10)}{Mouse}
}
}
\usage{
genePosition
}
\description{
A dataset containing the chromosomal coordinate for known human
and mouse genes
}
\keyword{datasets}
