% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{generateChromosomeGeneSet}
\alias{generateChromosomeGeneSet}
\title{Topological gene sets}
\usage{
generateChromosomeGeneSet(species = c("human", "mouse"), k = 100,
  skip = 25)
}
\arguments{
\item{species}{Character string indicating the species, either human or mouse}

\item{k}{Integer indicating the number of genes per set}

\item{skip}{Interger indicating the displacement of the window for selecting the k genes}
}
\value{
List of topoligically-close gene sets
}
\description{
This function generates a list of sets of k genes encoded by neighbor loci
}
