% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{getElementsFromString}
\alias{getElementsFromString}
\title{Get elements from string}
\usage{
getElementsFromString(x, sep = "-", pos = 1)
}
\arguments{
\item{x}{Vector of character strings}

\item{sep}{Character string indicating the separation character/s}

\item{pos}{Vector of integers indicating the positions of interest}
}
\value{
Vector or Matrix with same number of rows as elements in the input x vector and columns as positions selected
}
\description{
This function select the selected positions from a delimited string and return them as a matrix
}
