% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixGaussianFit.r
\name{getPeaks}
\alias{getPeaks}
\title{Find the modes for a multimodal distribution}
\usage{
getPeaks(x, adj = 1.5)
}
\arguments{
\item{x}{Numeric vector}

\item{adj}{Number indicating the adjust parameter for bandwidth of the density estimation}
}
\value{
Numeric vector of modes
}
\description{
This function returns the modes of a multimodal distribution
}
