% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixGaussianFit.r
\name{getPeaks3}
\alias{getPeaks3}
\title{Find the modes for a multimodal distribution}
\usage{
getPeaks3(x, adj = 1.2, thr = 0.01)
}
\arguments{
\item{x}{Numeric vector}

\item{adj}{Number indicating the adjust parameter for bandwidth of the density estimation}

\item{thr}{Threshold for lambda, distrivutions with lambda below this threshold will be discarded}
}
\value{
list of three elements: mean, sd and lambda
}
\description{
This function returns the modes of a multimodal distribution
}
