% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{hotSpotGenes}
\alias{hotSpotGenes}
\title{Select hotspot genes based on mean squared CNA values
This function selects hotspot genes based on mean squared CNA values}
\usage{
hotSpotGenes(nes, quantile)
}
\arguments{
\item{nes}{Matrix of inferred CNV scores}

\item{quantile}{Numeric value between 0 and 1 indicating the quantile to
calculate GI-scores including only top / "hotspot" genes according to their
squared inferred CNV values across all cells.}
}
\value{
Character vector of gene names corresponding to the selected hotspot genes
}
\description{
Select hotspot genes based on mean squared CNA values
This function selects hotspot genes based on mean squared CNA values
}
