% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{keepVectorElements}
\alias{keepVectorElements}
\title{This function filter a vector keeping only selected elements}
\usage{
keepVectorElements(x, elements)
}
\arguments{
\item{elements}{vector of elements to keep}
}
\value{
Filtered vector
}
\description{
This function filter a vector keeping only selected elements
}
