% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{matrixOrderRowsByName}
\alias{matrixOrderRowsByName}
\title{Matrix order rows by name}
\usage{
matrixOrderRowsByName(x, names)
}
\arguments{
\item{x}{Matrix}

\item{names}{Vector of strings}
}
\value{
Matrix with ordered rows
}
\description{
Matrix order rows by name
}
