% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixGaussianFit.r
\name{mixGaussianFit}
\alias{mixGaussianFit}
\title{Fit a mixture of gaussian curves}
\usage{
mixGaussianFit(x, thr = 0.01, min = 1, max = 1e+06, adj = c(0.8,
  2))
}
\arguments{
\item{x}{Numeric vectors}

\item{thr}{Minimum lambda (proportion of the distribution) to include in the analysis}

\item{min}{Optional minimum number of distributions}

\item{max}{Optional maximum number of distributions}

\item{adj}{Optional vector of 2 components defining the range of values for the density adj parameter}
}
\value{
Object of class mgfit. List of fitted parameters
}
\description{
This function fits a mixture of gaussians to the distribution of any population
}
