% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{msq}
\alias{msq}
\title{Mean squared method for genomic instability score}
\usage{
msq(nes, quantile = NULL, ref = NULL)
}
\arguments{
\item{nes}{Matrix of inferred CNV scores}

\item{quantile}{Optional number between 0 and 1 indicating the quantile to
calculate GI-scores including only top / "hotspot" genes according to their
squared inferred CNV values across all cells.}

\item{ref}{Optional character vector of cell names to be used as reference for
the selection of the top genes. If NULL, all cells will be used for the selection of the top genes.}
}
\value{
Numeric vector with the genomic instability score for each cell based on the mean squared method
}
\description{
This function computes the mean squared method for genomic instability score
}
