% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{nesTransformationFunction}
\alias{nesTransformationFunction}
\title{Transformation function for NES}
\usage{
nesTransformationFunction(x, method = c("scale", "equalize"))
}
\arguments{
\item{x}{Numeric vector}

\item{method}{Character string indicating the transformation method}
}
\value{
Function that takes a vector or matrix of values to be transformed and returns a vector or matrix of transformed values
}
\description{
Transformation function for NES
}
