% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{normalOverlap}
\alias{normalOverlap}
\title{Maximum overlap between reference and test gaussians}
\usage{
normalOverlap(mu1, mu2, sd1 = 1, sd2 = 1, tol = 1e-12)
}
\arguments{
\item{mu1}{Numeric mean for the first normal}

\item{mu2}{Numeric mean for the second normal}

\item{sd1}{Numeric standard deviation for the first normal}

\item{sd2}{Numeric standard deviation for the second normal}

\item{tol}{Numeric tolerance for numeric checks}
}
\value{
Numeric overlap in [0,1]
}
\description{
Overlap between two normal distributions
}
\details{
Compute the overlap integral \eqn{\int \min(f_1(x), f_2(x))\,dx} where
\eqn{f_1,f_2} are normal densities with given means and sds.
}
