% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot.inferCNV}
\alias{plot.inferCNV}
\title{Plot chromosome map}
\usage{
\method{plot}{inferCNV}(inferCNV, output = NULL, threshold = 0.2,
  gamma = 1.5, resolution = 150)
}
\arguments{
\item{inferCNV}{Object of class inferCNV}

\item{output}{Optional output PDF file name (with extension)}

\item{threshold}{Likelihood threshold for identifying genomically inestable cells/samples, 0 disables this filter}

\item{gamma}{Number indicating the gamma transformation for the colors}

\item{resolution}{Integer indicating the ppi for the png and jpg output files}
}
\value{
Nothing, a plot is generated in the default output devise
}
\description{
This function generates a chromosomes map plot for the inferred CNVs
}
