% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plothm}
\alias{plothm}
\title{Plot heatmap}
\usage{
plothm(x, color = c("royalblue", "firebrick2"), gama = 1, cex = 1,
  grid = T, scale = F, scmax = 0, box = TRUE, ...)
}
\arguments{
\item{x}{Numerical matrix}

\item{color}{Two character strings vector describing the colors for the heatmap}

\item{gama}{Number, indicating the exponential transformation for the color scale}

\item{cex}{Number indicating the magnification factor for the labels}

\item{grid}{Logical, whether a grid should be ploted}

\item{scale}{Number between 0 and .9 indicating the proportion of vertical space used to draw the color scale}

\item{scmax}{Optional number indicating the maximum value to be allowed for the heatmap}

\item{box}{Logical, whether to draw a box around the plot}

\item{...}{Additional parameters to pass to the plot function}
}
\value{
Nothing, a heatmap is produced in the default output device
}
\description{
This function produce a heatmap plot from a numerical matrix
}
