% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixGaussianFit.r
\name{predict.mgfit}
\alias{predict.mgfit}
\title{Predict relative likelihood for mixGaussianFit This function computes the
relative likelihood based on parameters fitted with mixGaussianFit function}
\usage{
\method{predict}{mgfit}(fit, x, k = NULL)
}
\arguments{
\item{fit}{Object generated by mixGaussianFit function}

\item{x}{Numerical vector of observations}

\item{k}{Optional vector of integers indicating the distributions to use}
}
\value{
Matrix of relative likelihood, observations in rows and distributions
in columns
}
\description{
Predict relative likelihood for mixGaussianFit This function computes the
relative likelihood based on parameters fitted with mixGaussianFit function
}
