% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{rowVars}
\alias{rowVars}
\title{Variance of rows for arrays with NA values}
\usage{
rowVars(x)
}
\arguments{
\item{x}{Numeric matrix}
}
\value{
Vector with the variance by row results
}
\description{
This function computes the variance by rows ignoring NA values
}
