% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{rowWMeans}
\alias{rowWMeans}
\title{Weighted mean by rows}
\usage{
rowWMeans(x, w)
}
\arguments{
\item{x}{Sparse matrix}

\item{w}{Numeric vector of weights}
}
\value{
Vector of weighted means
}
\description{
Weighted mean by rows
}
