% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{rowWVars}
\alias{rowWVars}
\title{Weighted variance by rows}
\usage{
rowWVars(x, w)
}
\arguments{
\item{x}{Sparse matrix}

\item{w}{Numeric vector of weights}
}
\value{
Vector of weighted variances
}
\description{
Weighted variance by rows
}
