% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{sortRowsByHclust}
\alias{sortRowsByHclust}
\title{Sort the rows of a matrix based on hierarchical cluster analysis}
\usage{
sortRowsByHclust(x, metric = c("euclidean", "maximum", "manhattan",
  "canberra", "binary", "minkowski"), method = c("complete", "ward.D",
  "ward.D2", "single", "complete", "average", "mcquitty", "median",
  "centroid"))
}
\arguments{
\item{x}{Numeric matrix}

\item{metric}{Character string indicating the distance metric}

\item{method}{Character string indicating the method for the hierarchical cluster analysis}
}
\value{
Numerical matrix
}
\description{
Sort the rows of a matrix based on hierarchical cluster analysis
}
