% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{sourceGenomicInstability}
\alias{sourceGenomicInstability}
\title{Genomic instability functions for debugging}
\usage{
sourceGenomicInstability(path = "~/code/genomicInstability")
}
\arguments{
\item{path}{String indicating the path to the genomicInstability package}
}
\description{
Genomic instability functions for debugging
}
