% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{splitByChromosome}
\alias{splitByChromosome}
\title{split matrix by chromosome}
\usage{
splitByChromosome(x)
}
\arguments{
\item{x}{matrix of NES with cells or samples in rows and chromosome fragments in columns}
}
\value{
List of matrices
}
\description{
split matrix by chromosome
}
