% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{splitMatrixByColumns}
\alias{splitMatrixByColumns}
\title{Split Matrix By Columns}
\usage{
splitMatrixByColumns(x, n = 1000)
}
\arguments{
\item{x}{Matrix}

\item{n}{Integer indicating the maximum column number for the pieces of the matrix}
}
\value{
List of matrices
}
\description{
This function split a based on the number of columns and returns a list of matrices
}
