% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{splitVectorWindow}
\alias{splitVectorWindow}
\title{Split a vector based on a window with and displacement}
\usage{
splitVectorWindow(x, k = 200, skip = 50)
}
\arguments{
\item{x}{Vector to split}

\item{k}{Integer indicating the window width}

\item{skip}{Integer indicating the displacement}
}
\value{
List of vectors with width = k
}
\description{
This function split a vector into a list of vectors with a given window width and displacement
}
