% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{subsetMatrixByColumns}
\alias{subsetMatrixByColumns}
\title{Subset Matrix by columns}
\usage{
subsetMatrixByColumns(i, x)
}
\arguments{
\item{i}{Vector indicating the columns to select}

\item{x}{Matrix}
}
\value{
Matrix with selected columns i
}
\description{
Subset Matrix by columns
}
