% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{subsetVector}
\alias{subsetVector}
\title{Get positions from vector}
\usage{
subsetVector(x, pos = 1)
}
\arguments{
\item{x}{Vector}

\item{pos}{Integer indicating the positios to return}
}
\value{
Verctor
}
\description{
This function returns a slected subset of a vector
}
