% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{subsetVectorRange}
\alias{subsetVectorRange}
\title{Range of elements from vector}
\usage{
subsetVectorRange(start_pos, x, k = 200)
}
\arguments{
\item{start_pos}{Integer indicating the starting position for the subset}

\item{x}{Vector}

\item{k}{Integer indicating the number of elements to include in the subset}
}
\value{
Vector
}
\description{
This function returns a subset of k elements of a vector starting at position start_pos
}
