% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{transformNesForPlot}
\alias{transformNesForPlot}
\title{Transform data for the heatmap plot}
\usage{
transformNesForPlot(x, method = c("scale", "equalize"))
}
\arguments{
\item{x}{either matrix or list of matrices}

\item{method}{Character string indicating the transformation method}
}
\value{
Transformed matrix or list of matrices
}
\description{
Transform data for the heatmap plot
}
