% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{validateInferCNV}
\alias{validateInferCNV}
\title{Validate genomicInstability objects}
\usage{
validateInferCNV(x, slots = "nes")
}
\arguments{
\item{x}{Instance of class inferCNV to check}

\item{slots}{String vector with slots to check}
}
\value{
x, error is trigered if the test is not suscessful
}
\description{
This function asserts the validity of inferCNV-class objects
}
