% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{genomicInstabilityScore}
\alias{genomicInstabilityScore}
\title{Genomic Instability Analysis}
\usage{
genomicInstabilityScore(cnv, likelihood = TRUE)
}
\arguments{
\item{cnv}{Object of class inferCNV generated by inferCNV() function}

\item{likelihood}{Logical, whether the genomic instability likelihood should be estimated}
}
\value{
Object of class inferCNV with updated slots for gis and gisnull
}
\description{
This function computes the genomic instability for an object of class inferCNV
}
