% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{giDensityPlot}
\alias{giDensityPlot}
\title{Genomic instability plot}
\usage{
giDensityPlot(inferCNV, legend = c("topleft", "top", "topright", "none"),
  ...)
}
\arguments{
\item{inferCNV}{Object of class inferCNV}

\item{legend}{Character string indicating the location of the legend. none to not include it}

\item{...}{Additional parameters for plot()}
}
\value{
None, a figure is created in the default output device
}
\description{
This function plot the genomic instability distribution, gaussian fits and null distribution if available
}
