% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferCNVsc}
\alias{inferCNVsc}
\title{Inference of CNV from single-cell sparse raw-count matices}
\usage{
inferCNVsc(
  expmat,
  nullmat = NULL,
  species = c("human", "mouse"),
  k = 100,
  skip = 25,
  min_geneset = 10,
  center = TRUE,
  scale = FALSE,
  cell_class = NULL,
  batch_size = 1000,
  verbose = TRUE
)
}
\arguments{
\item{expmat}{Sparse matrix of raw-counts}

\item{nullmat}{Optional sparse matrix with same number of rows as
\code{expmat} to be used as null model or index for the cells in expmat
considered a diploid reference}

\item{species}{Character string indicating the species, either human or mouse}

\item{k}{Integer indicating the number of genes per set}

\item{skip}{Interger indicating the displacement of the window for selecting
the k genes}

\item{min_geneset}{Integer indicating the minimum size for the genesets}

\item{center}{Logical, whether to subtract the mean expression of each gene}

\item{scale}{Logical, whether to divide by each gene standard deviation}

\item{cell_class}{Optional named vector of character strings indicating the
class of each cell in the expmat and nullmat matrices.
This is used for regularizing the mean and standard deviation estimation when
center and scale are TRUE, by using the mean and standard deviation of the
cells in the same class instead of all the cells. This is useful when there
are different cell types in the dataset with different expression profiles,
which can affect the estimation of the mean and standard deviation for each
gene.}

\item{batch_size}{Number indicating the maximum ize of the batch of cells
to be analyzed by aREA per iteration. This parameter is used to limit
memory load when transforming the sparse matrix normalized expression into
full matrix ranked signature}

\item{verbose}{Logical, whether progress should be reported}
}
\value{
Object of class inferCNV, which is a list containing matrix of nes,
and parameters (param), including species, window (k) and skip
}
\description{
This function estimates the CNV score based on single-cell expression data
}
\examples{
eh <- ExperimentHub::ExperimentHub()
dset <- eh[["EH5419"]]
tpm_matrix <- SummarizedExperiment::assays(dset)$TPM
set.seed(1)
tpm_matrix <- tpm_matrix[, sample(ncol(tpm_matrix), 500)]
cnv <- inferCNV(tpm_matrix)
class(cnv)
names(cnv)
cnv$nes[1:5, 1:3]

}
