% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixGaussianFit.r
\name{plot.mgfit}
\alias{plot.mgfit}
\title{Plot mixGaissuinFit objects}
\usage{
\method{plot}{mgfit}(fit, x, col = "grey75", lwd = 2, fitCol = NULL,
  main = "", ylim = NULL, ...)
}
\arguments{
\item{fit}{Object generated by the function mixGaussianFit}

\item{x}{Vector of values to plot the distribution}

\item{col}{Color for the background distribution}

\item{lwd}{Line weight for the fitted distributions}

\item{fitCol}{Optional vector of colors for the fitted distributions}

\item{main}{Character string for the main title}

\item{...}{Additional parameters to pass to the plot function}
}
\value{
Nothing, a plot is generated
}
\description{
Plot a mixture of distributions fitted with mixGaussianFit
}
